feat: Tidy existing loader hook
