+++ packages/node-core/src/sdk/esmLoader.ts
@@ -1,31 +1,31 @@
-import { consoleSandbox, debug, GLOBAL_OBJ } from '@sentry/core';
+import { debug, GLOBAL_OBJ } from '@sentry/core';
 import { createAddHookMessageChannel } from 'import-in-the-middle';
-import moduleModule from 'module';
+import * as moduleModule from 'module';
+import { supportsEsmLoaderHooks } from '../utils/detection';
 
-/** Initialize the ESM loader. */
-export function maybeInitializeEsmLoader(): void {
-  const [nodeMajor = 0, nodeMinor = 0] = process.versions.node.split('.').map(Number);
+/**
+ * Initialize the ESM loader - This method is private and not part of the public
+ * API.
+ *
+ * @ignore
+ */
+export function initializeEsmLoader(): void {
+  if (!supportsEsmLoaderHooks()) {
+    return;
+  }
 
-  // Register hook was added in v20.6.0 and v18.19.0
-  if (nodeMajor >= 21 || (nodeMajor === 20 && nodeMinor >= 6) || (nodeMajor === 18 && nodeMinor >= 19)) {
-    if (!GLOBAL_OBJ._sentryEsmLoaderHookRegistered) {
-      try {
-        const { addHookMessagePort } = createAddHookMessageChannel();
-        // @ts-expect-error register is available in these versions
-        moduleModule.register('import-in-the-middle/hook.mjs', import.meta.url, {
-          data: { addHookMessagePort, include: [] },
-          transferList: [addHookMessagePort],
-        });
-      } catch (error) {
-        debug.warn('Failed to register ESM hook', error);
-      }
+  if (!GLOBAL_OBJ._sentryEsmLoaderHookRegistered) {
+    GLOBAL_OBJ._sentryEsmLoaderHookRegistered = true;
+
+    try {
+      const { addHookMessagePort } = createAddHookMessageChannel();
+      // @ts-expect-error register is available in these versions
+      moduleModule.register('import-in-the-middle/hook.mjs', import.meta.url, {
+        data: { addHookMessagePort, include: [] },
+        transferList: [addHookMessagePort],
+      });
+    } catch (error) {
+      debug.warn("Failed to register 'import-in-the-middle' hook", error);
     }
-  } else {
-    consoleSandbox(() => {
-      // eslint-disable-next-line no-console
-      console.warn(
-        `[Sentry] You are using Node.js v${process.versions.node} in ESM mode ("import syntax"). The Sentry Node.js SDK is not compatible with ESM in Node.js versions before 18.19.0 or before 20.6.0. Please either build your application with CommonJS ("require() syntax"), or upgrade your Node.js version.`,
-      );
-    });
   }
 }
