--- packages/node-core/src/utils/commonjs.ts
@@ -1,8 +1,0 @@
-/** Detect CommonJS. */
-export function isCjs(): boolean {
-  try {
-    return typeof module !== 'undefined' && typeof module.exports !== 'undefined';
-  } catch {
-    return false;
-  }
-}
