+++ packages/node-core/src/utils/createMissingInstrumentationContext.ts
@@ -1,5 +1,5 @@
 import type { MissingInstrumentationContext } from '@sentry/core';
-import { isCjs } from './commonjs';
+import { isCjs } from './detection';
 
 export const createMissingInstrumentationContext = (pkg: string): MissingInstrumentationContext => ({
   package: pkg,
