+++ packages/node-core/src/utils/ensureIsWrapped.ts
@@ -1,8 +1,8 @@
 import { isWrapped } from '@opentelemetry/instrumentation';
 import { consoleSandbox, getClient, getGlobalScope, hasSpansEnabled, isEnabled } from '@sentry/core';
 import type { NodeClient } from '../sdk/client';
-import { isCjs } from './commonjs';
 import { createMissingInstrumentationContext } from './createMissingInstrumentationContext';
+import { isCjs } from './detection';
 
 /**
  * Checks and warns if a framework isn't wrapped by opentelemetry.
